[Normalization] Added sorting after inference of periods (Recidiviz/recidiviz-data#29599)
## Description of the change

recently this
[PR](https://github.com/Recidiviz/recidiviz-data/pull/29379/files) was
merged that, amongst other things, infers additional temporary IPs
whenever we see a SP with `supervision_level ==
StateSupervisionLevel.IN_CUSTODY in US_IX`. After merging though, we
realized that this change caused the normalization pipeline to crash
with a `Found open incarceration period with id [] preceding another
incarceration period with id []` error.

Long story short, the IPs are sorted before the IPs are inferred which
caused a situation where a closed inferred IP came after an open IP. So
this PR adds a line so that IPs are sorted before and after adding
inferred periods

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8e8637e824b071430c39d7c28c90ece3038bb706