[raw data validations] shrink stale table window from +2 days to + 12 hours (Recidiviz/recidiviz-data#29648)
## Description of the change

Per IE request, this PR shrinks the window for declaring a raw data
table as "stale" from `update_cadence_in_days` + 2 days to
`update_cadence_in_days` + 12 hours. We convert the time diff to hours
to apply the stale filter and then convert it back to days to make it
human readable in the validation output.

See
[here](https://console.cloud.google.com/bigquery?orgonly=true&project=recidiviz-123&supportedpurview=organizationId&ws=!1m4!1m3!3m2!1srecidiviz-123!2sethan_stale_test_validation_views)
for sandbox output.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29647

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5bd31091394ba2295a4975066c00e43a5b27ade0