[raw data import dag] adds new raw file metadata tables (Recidiviz/recidiviz-data#29525)
## Description of the change

See doc from [this
message](https://recidiviz.slack.com/archives/C028X32LRH7/p1714166212648459)
that explains the changes/motivations/questions around this schema
design

This PR adds two new tables `DirectIngestRawBigQueryFileMetadata` and
`DirectIngestRawGCSFileMetadata` and the necessary bits and bobs around
them. To reduce the size of the PR, the manager for these tables was
split into a different PR, see Recidiviz/recidiviz-data#29526.

I opted for `file_id` and `gcs_file_id` so that the `file_id` here would
match the `file_id` in our raw data tables; however, as noted in the
notion doc, I think updating the names to `file_id` and `batch_id` could
be clearer in the long term.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29443

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1250aa4314e26fea35a94d6461eb56aa1e8cd7fa