[raw data import dag] add raw file metadata manager v2 (Recidiviz/recidiviz-data#29526)
## Description of the change

Building off the work in Recidiviz/recidiviz-data#29525 , this PR adds a single metadata
manager, `DirectIngestRawFileMetadataManagerV2` to mediate interactions
with `DirectIngestRawBigQueryFileMetadata` and
`DirectIngestRawGCSFileMetadata`.

A few things to note --
- I'm not totally sure I approached schema -> entity conversion in the
right way, but in order to use the `OperationsSchemaToEntityConverter` I
added to an `operations_direction_checker` to allow us to populate
entity relationship.
- I went back and forth a little about how much to include `chunked`
file naming in the metadata manager -- opted for just the one kwarg
usage but am wondering if the over-abstraction of raw gcs/bq files
instead of chunked/not-chunk is not useful.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29443

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7d55f306412ae12b1fbd90b29ac26f07807e7a4f