add ME portion needed date to resident export (Recidiviz/recidiviz-data#29767)
This adds the portion served date to the resident record Firestore
export, since we want to display this for everyone in the JII webapp
pilot. There is an existing workaround for this in Workflows that
involved recalculating this date on the frontend, but I didn't want to
introduce another redundant calculation for the JII frontend.

GitOrigin-RevId: e6c8d867ef8e279df6c22da691430dada4e3abff