[US_CA] Add IDs to location_metadata (Recidiviz/recidiviz-data#29577)
## Description of the change

Several issues were occurring which have the same or partially the same
solution:
1. supervision_office_metrics_officer_assignment_sessions_materialized
is null for CA. This was because our district and office ids are null in
supervision_staff_attribute_sessions_materialized:
```
select *
from sessions.supervision_staff_attribute_sessions_materialized
where state_code = 'US_CA'
and supervision_office_id is not null and supervision_district_id is not null;
```
Returns no results.
This matters because in the assign CTE of
supervision_office_metrics_officer_assignment_sessions we have a final
where clause which says:
```
WHERE
        CONCAT(state_code, district, office) IS NOT NULL
```
and since district / office are null, the whole concat is null. District
and office above are aliases for the district and office IDs that I'm
adding to location metadata in this PR. [More context
here](https://recidiviz.slack.com/archives/C04G37PRQ10/p1714686372761419).
2. These validations (Recidiviz/recidiviz-data#29412, Recidiviz/recidiviz-data#29490) were failing because they expected
IDs

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29412, Recidiviz/recidiviz-data#29490

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cb45c8c48f1599c274fcaa08766ef856c8a9bcb0