[Insights] Maintain single supervisor and array on SupervisionOfficer in insights/schema.py (Recidiviz/recidiviz-data#29790)
## Description of the change

Add supervisor_external_ids array to the SupervisionOfficer entity in
the insights/schema.py.

In order to make this change without impacting the current data infra
used by the products, I created a separate BQ view that will output the
array and export this version to the Insights GCS bucket from the parent
branch of this PR. I'll clean up the diverging views as part of Recidiviz/recidiviz-data#29763.
I tested this by adding the array to fixture files and using the
`load_local_db` script. I then ran this query in postico to ensure that
the array column is imported and working as expected: `select
external_id, supervisor_id from supervision_officers,
unnest(supervisor_external_ids) as supervisor_id;`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7959c0719597098b6b204fe78caa86ca9685cd8f