[US_AR] Add AR location data to the location_metadata query (Recidiviz/recidiviz-data#29810)
## Description of the change

Location reference data has been created for AR and added to
reference_views (in Recidiviz/recidiviz-data#28452), but wasn't appended in the
location_metadata view itself. This change adds it to location_metadata,
which will allow AR location data to automatically start showing up in
incarceration_location_ids_to_names (due to the state agnostic bit at
the end of that query which pulls from location_metadata).

With this change in place, AR can be added to the location_record query
once the resident record query is merged. In doing this, it should first
be added to level_1_state_codes.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a69a03627213f9ee77a74a39d99d17be2b28c040