[Insights] Use Insights DB manager for ETL-related data (Recidiviz/recidiviz-data#29849)
## Description of the change

This PR instantiates a DB manager for the Insights DB in the
OutliersQuerier such that ETL-related data is queried from the Insights
DB whereas the non-ETL data (configurations and user_metadata tables)
are queried from the Outliers DB.

The PR is ready for review, however I'll keep in a draft state as a note
to myself to only make this PR merge-able after I test. I plan to test
the success of the new export/import for existing states by loading a
local DB with staging data then running a local Case Triage server and
using the `request_api` script to compare local results with staging
results. Testing is blocked on exporting views as JSON, which should
occur after the next deploy. I'll update this PR with comments once I
test after existing states (ie MI) have their results imported to the
staging DB.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Jen <jovergaag@gmail.com>
GitOrigin-RevId: 5aae2bb855822b4783a754dab149274b6cbc69ae