Add terraform to create ingest filename normalization cloud functions (Recidiviz/recidiviz-data#29652)
## Description of the change

For every primary and secondary ingest state bucket, create a cloud
function that listens for object finalization events (create/rename
file) in that bucket and calls `normalize_filename` function.

Some reformats as a result of running `terraform fmt`

For the file normalization function, a request takes 300-400MB memory,
so I’ve allocated 512MB to the function.
File unzipping is a longer process, taking ~2 minutes for a 700MB
zipfile. Since a cloud function times out after 9 minutes, the largest
zipfile we would be able to handle is ~3G. In testing the cloud function
needs roughly double the size of the zipfile in memory, so I chose 8G
for the unzipping function.

tfplan:
https://gist.github.com/emilyemilyemilyemilyemilyemily/df0bc2741221524c221569910cca54a6

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#29134

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0483738c9bde815943768135ec3ebd19f6eada0c