[US_AZ] Add raw data YAML for new validation data (Recidiviz/recidiviz-data#29949)
## Description of the change

Trying to set up an external comparison validation for AZ in Recidiviz/recidiviz-data#29938,
some CI tests are failing because they can't find the data used in the
validation view in any tracked source table location. I think that is
because the YAML for that source table is included in the same PR, so it
cannot be recognized by
`collect_all_source_table_configs._collect_raw_data_source_table_collections`,
since that table does not yet exist in BigQuery. This adds the YAML
separately so that tests will pass on the other PR.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8f25edbc4fe4ba998fe9476a3c93a258a9c95a8b