[US_CA] Include full agent classification in role period (Recidiviz/recidiviz-data#29868)
## Description of the change

We need to start including the full agent classification in the role
periods so that we can make sure PA II do not supervise PA III or other
PA II. This adds the classification to the role_subtype_raw_text, and
later we can use this to avoid creating supervisor periods we don't
want.

[Loom discussing the
change](https://www.loom.com/share/156246677ae84bf1b18cf9ebdbe30ac7) --
I don't know if this is actually helpful in this case, but trying it!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 87c4d74a4dd36e23680e55f06db359d2f9f87e81