[US_MO] Restrictive Housing - include assignments and sanctions in resident record (Recidiviz/recidiviz-data#29099)
## Description of the change

Right now, there is a sidebar element in the Restrictive Housing
workflows displaying a list of sanctions over the past year. This PR
does two things:
1. Adds a new element to the opportunity record showing a list of
assignments over the past year.
2. Adds a MO-specific resident metadata view including the list of
sanctions and the list of assignments. This is intended to support the
inclusion of these elements in the "all residents" view of the MO
opportunity, so that staff can view this information prior to someone's
assignment to Restrictive Housing.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28443
Closes Recidiviz/recidiviz-data#28442

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 069508994b10d0d453adc19970660b10f63e4f20